Cut Sticky from Header.jsx Somehow Semantics Sticky component fucks cy visit
